revised avatar action and reducer
